refactor: extract all class to main
